+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.3.4b -
 
-- Update Vercel CLI to 50.15.1
+- Update Vercel CLI to 50.15.1 and AWS-SDK
 
 ## 0.3.3b - 10/02/2026 19:10 UTC+1